+++ src/main/java/net/geforcemods/securitycraft/blocks/reinforced/ReinforcedWallBlock.java
@@ -2,6 +2,7 @@ package net.geforcemods.securitycraft.blocks.reinforced;
 
 import java.util.function.Supplier;
 
+import net.geforcemods.securitycraft.SCContent;
 import net.geforcemods.securitycraft.api.IReinforcedBlock;
 import net.geforcemods.securitycraft.api.OwnableBlockEntity;
 import net.geforcemods.securitycraft.misc.OwnershipEvent;
@@ -21,10 +22,13 @@ import net.neoforged.neoforge.common.NeoForge;
 public class ReinforcedWallBlock extends WallBlock implements IReinforcedBlock, EntityBlock {
 	private final Supplier<Block> vanillaBlockSupplier;
 
+	public ReinforcedWallBlock(Block vanillaBlock) {
+		this(SCContent.reinforcedCopy(vanillaBlock), vanillaBlock);
+	}
+
 	public ReinforcedWallBlock(BlockBehaviour.Properties properties, Block vanillaBlock) {
 		super(properties);
-
-		this.vanillaBlockSupplier = () -> vanillaBlock;
+		vanillaBlockSupplier = () -> vanillaBlock;
 	}
 
 	@Override